move arguments to subdirectory to enable fast imports of these
