+++ rasa_core/train.py
@@ -28,7 +28,7 @@ def create_argument_parser():
                     'your domain definition to train a dialogue '
                     'model to predict a bots actions.')
     parent_parser = argparse.ArgumentParser(add_help=False)
-    add_general_args(parent_parser)
+    cli.train.add_general_args(parent_parser)
 
     subparsers = parser.add_subparsers(
         help='Training mode of core.',
@@ -49,114 +49,11 @@ def create_argument_parser():
         help='teach the bot with interactive learning',
         parents=[parent_parser])
 
-    add_compare_args(compare_parser)
-    add_interactive_args(interactive_parser)
-    add_train_args(train_parser)
+    cli.train.add_compare_args(compare_parser)
+    cli.train.add_interactive_args(interactive_parser)
+    cli.train.add_train_args(train_parser)
 
     return parser
-
-
-def add_compare_args(parser):
-    parser.add_argument(
-        '--percentages',
-        nargs="*",
-        type=int,
-        default=[0, 5, 25, 50, 70, 90, 95],
-        help="Range of exclusion percentages")
-    parser.add_argument(
-        '--runs',
-        type=int,
-        default=3,
-        help="Number of runs for experiments")
-
-    cli.arguments.add_output_arg(
-        parser,
-        help_text="directory to persist the trained model in",
-        required=True)
-    cli.arguments.add_config_arg(
-        parser,
-        nargs="*")
-    cli.arguments.add_model_and_story_group(
-        parser,
-        allow_pretrained_model=False)
-    cli.arguments.add_domain_arg(
-        parser,
-        required=True)
-
-
-def add_interactive_args(parser):
-    parser.add_argument(
-        '-u', '--nlu',
-        type=str,
-        default=None,
-        help="trained nlu model")
-    parser.add_argument(
-        '--endpoints',
-        default=None,
-        help="Configuration file for the connectors as a yml file")
-    parser.add_argument(
-        '--skip_visualization',
-        default=False,
-        action='store_true',
-        help="disables plotting the visualization during "
-             "interactive learning")
-    parser.add_argument(
-        '--finetune',
-        default=False,
-        action='store_true',
-        help="retrain the model immediately based on feedback.")
-
-    cli.arguments.add_output_arg(
-        parser,
-        help_text="directory to persist the trained model in",
-        required=False)
-    cli.arguments.add_config_arg(
-        parser,
-        nargs=1)
-    cli.arguments.add_model_and_story_group(
-        parser,
-        allow_pretrained_model=True)
-    cli.arguments.add_domain_arg(
-        parser,
-        required=False)
-
-
-def add_train_args(parser):
-    cli.arguments.add_config_arg(
-        parser,
-        nargs=1)
-    cli.arguments.add_output_arg(
-        parser,
-        help_text="directory to persist the trained model in",
-        required=True)
-    cli.arguments.add_model_and_story_group(
-        parser,
-        allow_pretrained_model=False)
-    cli.arguments.add_domain_arg(
-        parser,
-        required=True)
-
-
-def add_general_args(parser):
-    parser.add_argument(
-        '--augmentation',
-        type=int,
-        default=50,
-        help="how much data augmentation to use during training")
-    parser.add_argument(
-        '--dump_stories',
-        default=False,
-        action='store_true',
-        help="If enabled, save flattened stories to a file")
-    parser.add_argument(
-        '--debug_plots',
-        default=False,
-        action='store_true',
-        help="If enabled, will create plots showing checkpoints "
-             "and their connections between story blocks in a  "
-             "file called `story_blocks_connections.html`.")
-
-    utils.add_logging_option_arguments(parser)
 
 
 def train_dialogue_model(domain_file, stories_file, output_path,
