+++ rasa_core/visualize.py
@@ -2,6 +2,7 @@ import argparse
 import logging
 import os
 
+import rasa_core.cli.arguments
 from rasa_core import utils, config, cli
 from rasa_core.agent import Agent
 
@@ -10,33 +11,13 @@ logger = logging.getLogger(__name__)
 
 def add_arguments(parser):
     """Parse all the command line arguments for the visualisation script."""
-    utils.add_logging_option_arguments(parser)
-    add_visualization_arguments(parser)
+    rasa_core.cli.arguments.add_logging_option_arguments(parser)
+    cli.visualization.add_visualization_arguments(parser)
     cli.arguments.add_config_arg(parser, nargs=1)
     cli.arguments.add_domain_arg(parser)
     cli.arguments.add_model_and_story_group(parser,
                                             allow_pretrained_model=False)
     return parser
-
-
-def add_visualization_arguments(parser):
-    parser.add_argument(
-        '-o', '--output',
-        default="graph.html",
-        type=str,
-        help="filename of the output path, e.g. 'graph.html")
-    parser.add_argument(
-        '-m', '--max_history',
-        default=2,
-        type=int,
-        help="max history to consider when merging "
-             "paths in the output graph")
-    parser.add_argument(
-        '-nlu', '--nlu_data',
-        default=None,
-        type=str,
-        help="path of the Rasa NLU training data, "
-             "used to insert example messages into the graph")
 
 
 def visualize(args):
